server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pgvector>=0.2.1
python-multipart>=0.0.6
pydantic>=2.0
speechbrain>=0.5
torch>=2.0
torchaudio>=2.0
numpy>=1.24
PyJWT>=2.8